fix(tablecard): allow formatted timestamps in expanded rows
