+++ src/index.js
@@ -72,9 +72,7 @@ export PieChartCard from './components/PieChartCard/PieChartCard';
 export ValueCard from './components/ValueCard/ValueCard';
 export TimeSeriesCard from './components/TimeSeriesCard/TimeSeriesCard';
 export ImageCard from './components/ImageCard/ImageCard';
-export TableCard, {
-  findMatchingThresholds,
-} from './components/TableCard/TableCard';
+export TableCard from './components/TableCard/TableCard';
 export GaugeCard from './components/GaugeCard/GaugeCard';
 export {
   CARD_TYPES,
@@ -306,6 +304,7 @@ export {
   getVariables,
   getCardVariables,
   replaceVariables,
+  findMatchingThresholds,
 } from './utils/cardUtilityFunctions';
 
 export { csvDownloadHandler } from './utils/componentUtilityFunctions';
