+++ src/utils/__tests__/cardUtilityFunctions.test.jsx
@@ -13,6 +13,7 @@ import {
   replaceVariables,
   chartValueFormatter,
   increaseSmallCardSize,
+  findMatchingThresholds,
 } from '../cardUtilityFunctions';
 import { CARD_SIZES } from '../../constants/LayoutConstants';
 
@@ -1176,6 +1177,93 @@ describe('cardUtilityFunctions', () => {
       expect(increaseSmallCardSize(CARD_SIZES.LARGEWIDE)).toEqual(
         CARD_SIZES.LARGEWIDE
       );
+    });
+  });
+  describe('findMatchingThresholds', () => {
+    const thresholds = [
+      { comparison: '>', dataSourceId: 'airflow_mean', severity: 3, value: 2 },
+      {
+        comparison: '>',
+        dataSourceId: 'airflow_mean',
+        severity: 1,
+        value: 2.2,
+      },
+      { comparison: '>', dataSourceId: 'airflow_max', severity: 3, value: 4 },
+      { comparison: '>', dataSourceId: 'airflow_max', severity: 1, value: 4.5 },
+      {
+        comparison: '=',
+        dataSourceId: 'airflow_status',
+        severity: 1,
+        value: 'High',
+      },
+    ];
+
+    it('findMatchingThresholds', () => {
+      const oneMatchingThreshold = findMatchingThresholds(
+        thresholds,
+        { airflow_mean: 4 },
+        'airflow_mean'
+      );
+      expect(oneMatchingThreshold).toHaveLength(1);
+      // The highest severity should match
+      expect(oneMatchingThreshold[0].severity).toEqual(1);
+    });
+    it('string value', () => {
+      const oneMatchingThreshold = findMatchingThresholds(
+        thresholds,
+        { airflow_status: 'High' },
+        'airflow_status'
+      );
+      expect(oneMatchingThreshold).toHaveLength(1);
+      // The highest severity should match
+      expect(oneMatchingThreshold[0].severity).toEqual(1);
+    });
+    it('multiple columns', () => {
+      const twoMatchingThresholds = findMatchingThresholds(thresholds, {
+        airflow_mean: 4,
+        airflow_max: 5,
+      });
+      expect(twoMatchingThresholds).toHaveLength(2);
+      // The highest severity should match
+      expect(twoMatchingThresholds[0].severity).toEqual(1);
+      expect(twoMatchingThresholds[0].dataSourceId).toEqual('airflow_mean');
+      expect(twoMatchingThresholds[1].severity).toEqual(1);
+      expect(twoMatchingThresholds[1].dataSourceId).toEqual('airflow_max');
+    });
+    it('no column', () => {
+      const thresholds = [
+        {
+          comparison: '<',
+          dataSourceId: 'airflow_mean',
+          severity: 1,
+          value: 4.5,
+        },
+      ];
+      const oneMatchingThreshold = findMatchingThresholds(thresholds, {
+        airflow_mean: 4,
+      });
+      expect(oneMatchingThreshold).toHaveLength(1);
+      // The highest severity should match
+      expect(oneMatchingThreshold[0].severity).toEqual(1);
+
+      // shouldn't match on null values
+      const zeroMatchingThreshold = findMatchingThresholds(thresholds, {
+        airflow_mean: null,
+      });
+      expect(zeroMatchingThreshold).toHaveLength(0);
+      // shouldn't match on null values
+      const zeroMatchingThreshold2 = findMatchingThresholds(
+        [
+          {
+            comparison: '<=',
+            dataSourceId: 'airflow_mean',
+            severity: 1,
+            value: 4.5,
+          },
+        ],
+        { airflow_mean: null }
+      );
+      expect(zeroMatchingThreshold2).toHaveLength(0);
     });
   });
 });
